working on rbshared

+++ matrix/rbshared/src/main/java/org/trifort/matrix/rbshared/MatrixRBShared.java
@@ -12,7 +12,7 @@ import org.trifort.rootbeer.runtime.Rootbeer;
 
 public class MatrixRBShared {
 
-  private void clearArray(double[][] array){
+  private void clearArray(float[][] array){
     for(int i = 0; i < array.length; ++i){
       for(int j = 0; j < array[0].length; ++j){
         array[i][j] = 0;
@@ -22,9 +22,9 @@ public class MatrixRBShared {
   
   public void run(){
     int size = MatrixGold.size();
-    double[][] a = MatrixGold.createAB();
-    double[][] b = MatrixGold.createAB();
-    double[][] c = new double[size][size];
+    float[][] a = MatrixGold.createAB();
+    float[][] b = MatrixGold.createAB();
+    float[][] c = new float[size][size];
     
     int blockCountX = 64;
     int blockCountY = 64;
